optimized. CLI is functional with api_methods rubyfile
