+++ lib/api_methods.rb
@@ -1,17 +1,46 @@
-def search_by_artist(artist)
-  api_result = RestClient::Request.execute(method: :get,
-      url: "https://api.harvardartmuseums.org/person",
-      headers: {params: {size: 100,
-                         q: "displayname:'#{artist}'",
-                         sort: "displayname",
-                         sortorder: "asc",
-                         fields: "displayname",
-                         apikey: ENV['API_KEY']}},)
-response_hash = JSON.parse(api_result)["records"]
+##### ARTWORK METHODS ######
+
+  def print_artwork_results(arr)
+    arr.each.with_index(0) do |data, index|
+      puts "#{index}. " + data['title'][0..100]
+      puts "\n"
+    end
+  end
+
+  def get_image_artist_name(arr, i)
+    puts "Artist: " + arr[i]["people"][0]["name"]
+  end
+
+  def get_image_url(arr, i)
+    puts "Image: " + arr[i]["primaryimageurl"]
+  end
+
+  def get_image_title(arr, i)
+   puts "Title: " + arr[i]["title"]
+  end
+
+  def open_image(arr, i)
+    imgurl = arr[i]["primaryimageurl"]
+    system("open -a Safari #{imgurl}")
+  end
+
+########### ARTIST METHODS #############
+
+def print_artist_results(arr)
+  arr.each.with_index(0) do |arr, index|
+    puts "#{index}. " + arr['displayname']
+    puts "\n"
+  end
 end
 
-# search_by_artist("magritte")
+def get_artist_name(arr, i)
+  arr[i]["displayname"]
+end
 
-# binding.pry
+def get_artist_culture(arr, i)
+  arr[i]["culture"]
+end
 
-0
+def get_artist_url(arr, i)
+  arr[i]["url"]
+end
